+++ Packages/Syntax/Sources/Syntax/RegexParser/RegexHighlightParser.swift
@@ -38,14 +38,20 @@ actor RegexHighlightParser: HighlightParsing {
     
     private let extractors: [SyntaxType: [any HighlightExtractable]]
     private let nestables: [NestableToken: SyntaxType]
+    private let lexicalRules: Syntax.LexicalRules
     
     
     // MARK: Lifecycle
     
-    init(extractors: [SyntaxType: [any HighlightExtractable]], nestables: [NestableToken: SyntaxType]) {
+    init(
+        extractors: [SyntaxType: [any HighlightExtractable]],
+        nestables: [NestableToken: SyntaxType],
+        lexicalRules: Syntax.LexicalRules = .default
+    ) {
         
         self.extractors = extractors
         self.nestables = nestables
+        self.lexicalRules = lexicalRules
     }
     
     
@@ -60,8 +66,8 @@ actor RegexHighlightParser: HighlightParsing {
     /// - Throws: `CancellationError`.
     func parseHighlights(in string: String, range: NSRange) async throws -> (highlights: [Highlight], updateRange: NSRange)? {
         
-        try await withThrowingTaskGroup { [extractors, nestables] group in
-            group.addTask { try nestables.parseHighlights(in: string, range: range) }
+        try await withThrowingTaskGroup { [extractors, nestables, lexicalRules] group in
+            group.addTask { try nestables.parseHighlights(in: string, range: range, delimiterEscapeRule: lexicalRules.delimiterEscapeRule) }
             
             for (type, extractors) in extractors {
                 for extractor in extractors {